Fixed products with , notation and confirm messages on dels

+++ resources/views/dashboard/category/overview.blade.php
@@ -28,7 +28,7 @@
                  <td>{{$category->desc}}</td>
                  <td>
                     <a class="btn btn-raised btn-warning btn-doggofood" href="{{ URL::to('dashboard/category/edit/' . $category->id) }}">Aanpassen</a>
-                    {{ Form::open(['id' => 'delete','method' => 'DELETE', 'route' => ['category_destroy', $category->id], 'style' => 'display: inline-block;']) }}
+                    {{ Form::open(['id' => 'delete','method' => 'DELETE', 'route' => ['category_destroy', $category->id], 'style' => 'display: inline-block;','onsubmit' => "return confirm('Record verwijderen?')"]) }}
                         {{ Form::submit('Verwijderen', ['class' => 'btn btn-raised btn-danger btn-doggofood']) }}
                     {{ Form::close() }}
                  </td>
@@ -57,7 +57,7 @@
                  <td>{{$subcategory->desc}}</td>
                  <td>
                     <a class="btn btn-raised btn-warning btn-doggofood" href="{{ URL::to('dashboard/subcategory/edit/' . $subcategory->id) }}">Aanpassen</a>
-                    {{ Form::open(['id' => 'delete','method' => 'DELETE', 'route' => ['subcategory_destroy', $subcategory->id], 'style' => 'display: inline-block;']) }}
+                    {{ Form::open(['id' => 'delete','method' => 'DELETE', 'route' => ['subcategory_destroy', $subcategory->id], 'style' => 'display: inline-block;','onsubmit' => "return confirm('Record verwijderen?')"]) }}
                         {{ Form::submit('Verwijderen', ['class' => 'btn btn-raised btn-danger btn-doggofood']) }}
                     {{ Form::close() }}
                  </td>
